+++ apps/web/src/app/(docs)/components/button/variants/sizes.tsx
@@ -3,7 +3,7 @@ import { Button, type ButtonProps } from "@jamsr-ui/react";
 export const ButtonSizes = () => {
   const sizes: ButtonProps["size"][] = ["sm", "md", "lg"];
   return (
-    <div className="flex flex-col gap-4">
+    <div className="flex items-center gap-4">
       {sizes.map((size) => (
         <Button key={size} size={size}>
           Button