+++ apps/web/src/app/(docs)/components/editor/page.tsx
@@ -1,7 +1,9 @@
 import { VariantPage } from "@/components/variant-page";
 import { VariantWrapper } from "@/components/variant-wrapper";
 import { type Metadata } from "next";
+import { EditorControlled } from "./variants/controlled";
 import { EditorDefault } from "./variants/default";
+import { EditorInvalid } from "./variants/invalid";
 
 export const metadata: Metadata = {
   title: "Editor",
@@ -12,6 +14,12 @@ const Editor = () => {
     <VariantPage heading="Editor">
       <VariantWrapper heading="Default">
         <EditorDefault />
+      </VariantWrapper>
+      <VariantWrapper heading="Controlled">
+        <EditorControlled />
+      </VariantWrapper>
+      <VariantWrapper heading="Invalid State">
+        <EditorInvalid />
       </VariantWrapper>
     </VariantPage>
   );