+++ apps/web/src/app/(docs)/components/editor/variants/controlled.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { Editor } from "@jamsr-ui/react";
+import { useState } from "react";
+
+export const EditorControlled = () => {
+  const [value, setValue] = useState({});
+  return (
+    <div>
+      <Editor value={value} onValueChange={setValue} />
+      {JSON.stringify(value, null, 2)}
+    </div>
+  );
+};