+++ apps/web/src/app/(docs)/components/editor/variants/invalid.tsx
@@ -1,0 +1,5 @@
+import { Editor } from "@jamsr-ui/react";
+
+export const EditorInvalid = () => {
+  return <Editor isInvalid helperText="Description is required" />;
+};