+++ packages/components/card/src/index.ts
@@ -4,4 +4,3 @@ export { Card, type CardProps } from "./card";
 export { CardContent, type CardContentProps } from "./card-content";
 export { CardFooter, type CardFooterProps } from "./card-footer";
 export { CardHeader, type CardHeaderProps } from "./card-header";
-