+++ packages/components/input/src/use-input.ts
@@ -176,7 +176,7 @@ export const useInput = (props: UseInputProps) => {
           className: cn(classNames?.base, props?.className),
         }),
         ref: baseRef,
-        "data-focus": dataAttr(isFocused),
+        "data-focused": dataAttr(isFocused),
         "data-filled-within": dataAttr(isFilledWithin),
         "data-has-label": dataAttr(hasLabel),
         "data-has-start-content": dataAttr(hasStartContent),