+++ packages/components/select/src/style.ts
@@ -14,7 +14,7 @@ export const selectVariant = tv({
     innerWrapper: "flex w-full items-center gap-2",
     label: "mb-1 text-sm text-foreground-400",
     trigger:
-      "relative flex w-full flex-row items-center gap-3 rounded-xl border-2 border-divider px-3 py-2 shadow-sm outline-none focus:!border-primary group-data-[hover=true]:border-gray-400 group-data-[open=true]:!border-primary",
+      "relative flex w-full flex-row items-center gap-3 rounded-xl border-2 border-divider px-3 py-2 shadow-sm outline-none focus:!border-primary group-data-[hovered=true]:border-gray-400 group-data-[open=true]:!border-primary",
     helperText: "mt-1 text-xs text-foreground-500",
     popover: "",
     indicator: