+++ packages/styles/src/styles-provider.tsx
@@ -1,5 +1,6 @@
 import { createContext, useContext } from "react";
 
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface UIStylesType {}
 const UIStyleContext = createContext<UIStylesType>(
   {} as unknown as UIStylesType,