+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/menu
 
+## 0.0.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.54
+
 ## 0.0.71
 
 ### Patch Changes