+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/popover
 
+## 0.0.58
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.54
+
 ## 0.0.57
 
 ### Patch Changes