+++ packages/components/progress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/progress
 
+## 0.0.57
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.54
+
 ## 0.0.56
 
 ### Patch Changes