+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/switch
 
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.54
+  - @jamsr-ui/typography@0.0.65
+
 ## 0.0.65
 
 ### Patch Changes