+++ packages/components/typography/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/typography
 
+## 0.0.65
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/utils@0.0.54
+
 ## 0.0.64
 
 ### Patch Changes