+++ packages/react/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @jamsr-ui/react
 
+## 0.0.156
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @jamsr-ui/file-upload-multi@0.0.79
+  - @jamsr-ui/textarea@0.0.72
+  - @jamsr-ui/button@0.0.60
+  - @jamsr-ui/dialog@0.0.67
+  - @jamsr-ui/select@0.0.73
+  - @jamsr-ui/input@0.0.72
+  - @jamsr-ui/table@0.0.37
+  - @jamsr-ui/card@0.0.69
+  - @jamsr-ui/styles@0.0.4
+  - @jamsr-ui/utils@0.0.54
+  - @jamsr-ui/rhf@0.0.9
+  - @jamsr-ui/alert@0.0.73
+  - @jamsr-ui/autocomplete@0.0.77
+  - @jamsr-ui/chip@0.0.65
+  - @jamsr-ui/confirmation@0.0.44
+  - @jamsr-ui/data-table@0.0.38
+  - @jamsr-ui/drawer@1.0.53
+  - @jamsr-ui/editor@0.0.70
+  - @jamsr-ui/header@0.0.63
+  - @jamsr-ui/menu@0.0.72
+  - @jamsr-ui/popover@0.0.58
+  - @jamsr-ui/tab@0.0.75
+  - @jamsr-ui/toast@0.0.20
+  - @jamsr-ui/tooltip@0.0.57
+  - @jamsr-ui/tags-input@0.0.8
+  - @jamsr-ui/radio@0.0.65
+  - @jamsr-ui/accordion@0.0.67
+  - @jamsr-ui/avatar@0.0.59
+  - @jamsr-ui/badge@0.0.57
+  - @jamsr-ui/checkbox@0.0.57
+  - @jamsr-ui/divider@0.0.58
+  - @jamsr-ui/file-upload@0.0.79
+  - @jamsr-ui/link@0.0.59
+  - @jamsr-ui/otp-input@0.0.57
+  - @jamsr-ui/progress@0.0.57
+  - @jamsr-ui/rating@0.0.16
+  - @jamsr-ui/ripple@0.0.58
+  - @jamsr-ui/skeleton@0.0.57
+  - @jamsr-ui/switch@0.0.66
+  - @jamsr-ui/typography@0.0.65
+
 ## 0.0.155
 
 ### Patch Changes