+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/styles
 
+## 0.0.4
+
+### Patch Changes
+
+- fix
+
 ## 0.0.3
 
 ### Patch Changes