+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/utils
 
+## 0.0.54
+
+### Patch Changes
+
+- fix
+
 ## 0.0.53
 
 ### Patch Changes